Address OAuth MCP review feedback
